// ux/api/client/types.ts
export interface components {
    schemas: {
// ...
        FindTermResult: {
            /** Title */
            title: string;
            /** Definition */
            definition: string;
            /**
             * Domain Id
             * Format: uuid
             */
            domain_id: string;
            /**
             * Layer Id
             * Format: uuid
             */
            layer_id: string;
            /** Parent Term Id */
            parent_term_id?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Title Embedding */
            title_embedding?: number[] | null;
            /** Definition Embedding */
            definition_embedding?: number[] | null;
            /** Created At */
            created_at: string;
            /** Version */
            version: number;
            /** Last Modified */
            last_modified: string;
            /** Score */
            score: number;
            /** Distance */
            distance: number;
        };
// ...
        TermCreate: {
            /** Title */
            title: string;
            /** Definition */
            definition: string;
            /**
             * Domain Id
             * Format: uuid
             */
            domain_id: string;
            /**
             * Layer Id
             * Format: uuid
             */
            layer_id: string;
            /** Parent Term Id */
            parent_term_id?: string | null;
        };
        /** TermOut */
        TermOut: {
            /** Title */
            title: string;
            /** Definition */
            definition: string;
            /**
             * Domain Id
             * Format: uuid
             */
            domain_id: string;
            /**
             * Layer Id
             * Format: uuid
             */
            layer_id: string;
            /** Parent Term Id */
            parent_term_id?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Title Embedding */
            title_embedding?: number[] | null;
            /** Definition Embedding */
            definition_embedding?: number[] | null;
            /** Created At */
            created_at: string;
            /** Version */
            version: number;
            /** Last Modified */
            last_modified: string;
        };
// ...
        TermUpdate: {
            /** Title */
            title?: string | null;
            /** Definition */
            definition?: string | null;
            /** Parent Term Id */
            parent_term_id?: string | null;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}